[benchmark] Move check-added to run_smoke_bench
